# contracts/openapi/diagnosis.openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "AgroDiagn\u00f3stico diagnosis",
    "version": "1.0.0",
    "description": "Contrato inicial: salud t\u00e9cnica implementada; operaciones de negocio contract-only. Sin servidores p\u00fablicos inventados."
  },
  "paths": {
...
    "/api/v1/diagnoses": {
      "post": {
        "operationId": "diagnosis_create",
        "summary": "Aceptar imagen para diagn\u00f3stico",
        "description": "Validar bytes/decodificaci\u00f3n antes de guardar y de publicar evento; imagen agron\u00f3micamente insuficiente se analiza y puede abstenerse. No exige cultivo. Misma clave+bytes en mismo propietario/operaci\u00f3n durante 86400 s devuelve ID/created_at originales y estado actual; contenido distinto 409 IDEMPOTENCY_CONFLICT. Plazo desde primera aceptaci\u00f3n, no deslizante: now >= expires_at permite nueva solicitud. Otra cuenta tiene namespace independiente. Fingerprint SHA-256 de bytes de imagen exactos (no filename, MIME declarado ni boundary); no hay otros campos sem\u00e1nticos admitidos. Borrado l\u00f3gico durante retenci\u00f3n: replay 409 IDEMPOTENCY_RESOURCE_DELETED, nunca resucita. Operaci\u00f3n concurrente a\u00fan en curso:409 IDEMPOTENCY_IN_PROGRESS, reintento posterior.",
        "security": [
          {
            "UserBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "required": false,
            "schema": {
              "$ref": "../schemas/common.schema.json#/$defs/UUID"
            },
            "example": "11111111-1111-4111-8111-111111111111",
            "description": "UUID del cliente o generado por servidor si falta; valor inv\u00e1lido 400; eco en respuesta."
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "pattern": "^[A-Za-z0-9._:-]+$"
            },
            "example": "carga-sintetica-001"
          }
        ],
        "responses": {
          "202": {
            "description": "Respuesta prevista; a\u00fan no implementada.",
            "headers": {
              "X-Correlation-ID": {
                "required": true,
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/UUID"
                },
                "example": "11111111-1111-4111-8111-111111111111"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "private, no-store"
                },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/DiagnosisAccepted"
                },
                "example": {
                  "id": "11111111-1111-4111-8111-111111111111",
                  "status": "PENDIENTE",
                  "created_at": "2026-09-28T12:00:00Z"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inv\u00e1lida.",
            "headers": {
              "X-Correlation-ID": {
                "required": true,
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/UUID"
                },
                "example": "11111111-1111-4111-8111-111111111111"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "private, no-store"
                },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/Error"
                },
                "example": {
                  "code": "INVALID_REQUEST",
                  "message": "Solicitud inv\u00e1lida.",
                  "correlation_id": "11111111-1111-4111-8111-111111111111"
                }
              }
            }
          },
          "401": {
            "description": "Autenticaci\u00f3n requerida o inv\u00e1lida.",
            "headers": {
              "X-Correlation-ID": {
                "required": true,
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/UUID"
                },
                "example": "11111111-1111-4111-8111-111111111111"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "private, no-store"
                },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/Error"
                },
                "example": {
                  "code": "UNAUTHORIZED",
                  "message": "Autenticaci\u00f3n requerida o inv\u00e1lida.",
                  "correlation_id": "11111111-1111-4111-8111-111111111111"
                }
              }
            }
          },
          "409": {
            "description": "La operaci\u00f3n entra en conflicto con el estado actual.",
            "headers": {
              "X-Correlation-ID": {
                "required": true,
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/UUID"
                },
                "example": "11111111-1111-4111-8111-111111111111"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "private, no-store"
                },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/Error"
                },
                "examples": {
                  "IDEMPOTENCY_CONFLICT": {
                    "value": {
                      "code": "IDEMPOTENCY_CONFLICT",
                      "message": "Conflicto con la clave de idempotencia.",
                      "correlation_id": "11111111-1111-4111-8111-111111111111"
                    }
                  },
                  "IDEMPOTENCY_RESOURCE_DELETED": {
                    "value": {
                      "code": "IDEMPOTENCY_RESOURCE_DELETED",
                      "message": "Conflicto con la clave de idempotencia.",
                      "correlation_id": "11111111-1111-4111-8111-111111111111"
                    }
                  },
                  "IDEMPOTENCY_IN_PROGRESS": {
                    "value": {
                      "code": "IDEMPOTENCY_IN_PROGRESS",
                      "message": "Conflicto con la clave de idempotencia.",
                      "correlation_id": "11111111-1111-4111-8111-111111111111"
                    }
                  }
                }
              }
            }
          },
          "413": {
            "description": "La imagen supera los l\u00edmites permitidos.",
            "headers": {
              "X-Correlation-ID": {
                "required": true,
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/UUID"
                },
                "example": "11111111-1111-4111-8111-111111111111"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "private, no-store"
                },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/Error"
                },
                "example": {
                  "code": "IMAGE_TOO_LARGE",
                  "message": "La imagen supera los l\u00edmites permitidos.",
                  "correlation_id": "11111111-1111-4111-8111-111111111111"
                }
              }
            }
          },
          "415": {
            "description": "Formato de imagen no admitido.",
            "headers": {
              "X-Correlation-ID": {
                "required": true,
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/UUID"
                },
                "example": "11111111-1111-4111-8111-111111111111"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "private, no-store"
                },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/Error"
                },
                "example": {
                  "code": "UNSUPPORTED_MEDIA_TYPE",
                  "message": "Formato de imagen no admitido.",
                  "correlation_id": "11111111-1111-4111-8111-111111111111"
                }
              }
            }
          },
          "503": {
            "description": "Fallo transitorio o indisponibilidad en almacenamiento o persistencia.",
            "headers": {
              "X-Correlation-ID": {
                "required": true,
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/UUID"
                },
                "example": "11111111-1111-4111-8111-111111111111"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "private, no-store"
                },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/Error"
                },
                "examples": {
                  "STORAGE_UNAVAILABLE": {
                    "value": {
                      "code": "STORAGE_UNAVAILABLE",
                      "message": "Almacenamiento temporalmente no disponible.",
                      "correlation_id": "11111111-1111-4111-8111-111111111111"
                    }
                  },
                  "PERSISTENCE_UNAVAILABLE": {
                    "value": {
                      "code": "PERSISTENCE_UNAVAILABLE",
                      "message": "Persistencia temporalmente no disponible.",
                      "correlation_id": "11111111-1111-4111-8111-111111111111"
                    }
                  }
                }
              }
            }
          }
        },
        "x-status": "implemented",
        "x-owner": "diagnosis",
        "x-increment": 2,
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Bytes de archivo, no base64; un \u00fanico campo obligatorio. M\u00e1ximo 10485760 bytes y 24000000 p\u00edxeles decodificados."
                  }
                },
                "required": [
                  "image"
                ],
                "additionalProperties": false
              },
              "encoding": {
                "image": {
                  "contentType": "image/jpeg, image/png, image/webp"
                }
              },
              "example": {
                "image": "REPRESENTACION_SINTETICA_NO_ES_UN_ARCHIVO"
              }
            }
          }
        },
        "x-upload-policy": {
          "max_file_bytes": 10485760,
          "max_decoded_pixels": 24000000,
          "supported_media_types": [
            "image/jpeg",
            "image/png",
            "image/webp"
          ],
          "v1_pending_media_types": [
            "image/heic",
            "image/heif"
          ],
          "invalid_decode_status": 400,
          "max_size_status": 413,
          "unsupported_status": 415,
          "e2e_verified": false
        },
        "x-idempotency": {
          "namespace": [
            "owner_id",
            "operation",
            "key"
          ],
          "retention_seconds": 86400,
          "expires_when": "now >= first_accepted_at + retention_seconds",
          "sliding": false,
          "same_content": "same-id-current-status",
          "different_content": 409,
          "different_owner": "independent",
          "expired": "new-request",
          "deleted": 409,
          "in_progress": 409
        }
      },
      "get": {
        "operationId": "diagnosis_list",
        "summary": "Historial propio paginado",
        "description": "Excluye borrado l\u00f3gico. Sin filtros adicionales en v1 inicial. Cursor inv\u00e1lido/ajeno 400 gen\u00e9rico, nunca revela propietario; orden created_at DESC, id DESC.",
        "security": [
          {
            "UserBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "required": false,
            "schema": {
              "$ref": "../schemas/common.schema.json#/$defs/UUID"
            },
            "example": "11111111-1111-4111-8111-111111111111",
            "description": "UUID del cliente o generado por servidor si falta; valor inv\u00e1lido 400; eco en respuesta."
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "example": 20
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 2048
            },
            "example": "cursor-opaco-de-ejemplo",
            "description": "Opaco, vinculado a propietario/filtros; inv\u00e1lido 400. Orden created_at DESC, id DESC; keyset, no offset."
          }
        ],
        "responses": {
          "200": {
            "description": "Respuesta prevista; a\u00fan no implementada.",
            "headers": {
              "X-Correlation-ID": {
                "required": true,
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/UUID"
                },
                "example": "11111111-1111-4111-8111-111111111111"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "private, no-store"
                },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/DiagnosisPage"
                },
                "example": {
                  "items": [
                    {
                      "id": "11111111-1111-4111-8111-111111111111",
                      "status": "PENDIENTE",
                      "created_at": "2026-09-28T12:00:00Z"
                    }
                  ],
                  "next_cursor": null
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inv\u00e1lida.",
            "headers": {
              "X-Correlation-ID": {
                "required": true,
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/UUID"
                },
                "example": "11111111-1111-4111-8111-111111111111"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "private, no-store"
                },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/Error"
                },
                "example": {
                  "code": "INVALID_REQUEST",
                  "message": "Solicitud inv\u00e1lida.",
                  "correlation_id": "11111111-1111-4111-8111-111111111111"
                }
              }
            }
          },
          "401": {
            "description": "Autenticaci\u00f3n requerida o inv\u00e1lida.",
            "headers": {
              "X-Correlation-ID": {
                "required": true,
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/UUID"
                },
                "example": "11111111-1111-4111-8111-111111111111"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "private, no-store"
                },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/Error"
                },
                "example": {
                  "code": "UNAUTHORIZED",
                  "message": "Autenticaci\u00f3n requerida o inv\u00e1lida.",
                  "correlation_id": "11111111-1111-4111-8111-111111111111"
                }
              }
            }
          }
        },
        "x-status": "implemented",
        "x-owner": "diagnosis",
        "x-increment": 2
      }
    },
...
    "/api/v1/admin/diagnoses": {
      "get": {
        "operationId": "diagnosis_admin_list_diagnoses",
        "summary": "Supervisi\u00f3n administrativa de metadatos de diagn\u00f3sticos",
        "description": "Solo rol ADMIN. Retorna listado paginado con cursor HMAC de diagn\u00f3sticos no borrados (id, owner_id, status, created_at, updated_at, reason_code). Excluye fotos, object_key, feedback y datos de Identity. Audita consulta de supervisi\u00f3n.",
        "security": [
          {
            "UserBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "example": 20
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 2048
            },
            "example": "cursor-admin-opaco",
            "description": "Cursor HMAC opaco para supervisi\u00f3n."
          },
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "required": false,
            "schema": {
              "$ref": "../schemas/common.schema.json#/$defs/UUID"
            },
            "example": "11111111-1111-4111-8111-111111111111",
            "description": "UUID del cliente o generado por servidor si falta; valor inv\u00e1lido 400; eco en respuesta."
          }
        ],
        "responses": {
          "200": {
            "description": "P\u00e1gina de diagn\u00f3sticos para supervisi\u00f3n.",
            "headers": {
              "X-Correlation-ID": {
                "required": true,
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/UUID"
                },
                "example": "11111111-1111-4111-8111-111111111111"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "private, no-store"
                },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/AdminDiagnosisPage"
                },
                "example": {
                  "items": [
                    {
                      "id": "11111111-1111-4111-8111-111111111111",
                      "owner_id": "33333333-3333-4333-8333-333333333333",
                      "status": "PENDIENTE",
                      "reason_code": null,
                      "created_at": "2026-09-30T12:00:00Z",
                      "updated_at": "2026-09-30T12:00:00Z"
                    }
                  ],
                  "next_cursor": null
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inv\u00e1lida o par\u00e1metros incorrectos.",
            "headers": {
              "X-Correlation-ID": {
                "required": true,
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/UUID"
                },
                "example": "11111111-1111-4111-8111-111111111111"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "private, no-store"
                },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/Error"
                },
                "examples": {
                  "INVALID_PAGINATION": {
                    "value": {
                      "code": "INVALID_PAGINATION",
                      "message": "Cursor HMAC inv\u00e1lido, alterado o par\u00e1metros inv\u00e1lidos.",
                      "correlation_id": "11111111-1111-4111-8111-111111111111"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Autenticaci\u00f3n requerida o token inv\u00e1lido.",
            "headers": {
              "X-Correlation-ID": {
                "required": true,
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/UUID"
                },
                "example": "11111111-1111-4111-8111-111111111111"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "private, no-store"
                },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/Error"
                },
                "example": {
                  "code": "UNAUTHORIZED",
                  "message": "Credenciales ausentes, vencidas o firma inv\u00e1lida.",
                  "correlation_id": "11111111-1111-4111-8111-111111111111"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente para ejecutar la operaci\u00f3n.",
            "headers": {
              "X-Correlation-ID": {
                "required": true,
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/UUID"
                },
                "example": "11111111-1111-4111-8111-111111111111"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "private, no-store"
                },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/Error"
                },
                "example": {
                  "code": "FORBIDDEN",
                  "message": "Se requiere rol administrativo para esta ruta.",
                  "correlation_id": "11111111-1111-4111-8111-111111111111"
                }
              }
            }
          }
        },
        "x-status": "implemented",
        "x-owner": "diagnosis",
        "x-increment": 2,
        "x-visibility": "admin"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "UserBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Firma, issuer, audience, expiry, revocaci\u00f3n/estado de cuenta. Solo identidad de usuario."
      },
      "ServiceBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Identidad ai_inference; audience interna dedicada. Bearer de usuario no autoriza. Algoritmo/configuraci\u00f3n en ADR de seguridad antes de incremento 1/3."
      },
      "RefreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Secure-agro_refresh",
        "description": "Cookie HttpOnly; Secure; SameSite=Lax; Path=/api/v1/auth; sin Domain. Rotaci\u00f3n y revocaci\u00f3n."
      },
      "CsrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Debe coincidir con token ligado a sesi\u00f3n; adem\u00e1s verificar Origin permitido."
      }
    }
  }
}